problems

+++ rgb/rgb7033.cpp
@@ -5,11 +5,11 @@ int main() {
     double a, b;
     scanf("%lf %lf", &a, &b);
 
-    int c, d, e, f;
-    scanf("%d %d %d %d", &c, &d, &e, &f);
+    double c, d, e, f;
+    scanf("%lf %lf %lf %lf", &c, &d, &e, &f);
 
-    int g, h, i, j, k, l;
-    scanf("%d %d %d %d %d %d", &g, &h, &i, &j, &k, &l);
+    double g, h, i, j, k, l;
+    scanf("%lf %lf %lf %lf %lf %lf", &g, &h, &i, &j, &k, &l);
 
     printf("%.3lf\n", abs(a-b));
     printf("%.3lf\n", pow(pow(abs(e-c), 2) + pow(abs(f-d), 2), 0.5));
